refactor: remove sensitive data from .env; add .env.example for configuration; implement category API structure with types and endpoints
