--- .env
@@ -1,1 +1,0 @@
-VITE_API_KEY='jWbpWSe3AJbIkDyneQqBO6xNP5rOlqd2hO8hhwug'
