+++ .env.example
@@ -1,0 +1,2 @@
+VITE_API_BASE_URL=https://quizapi.io/api/v1
+VITE_API_KEY=your_api_key_here
