+++ src/entities/Category/api/getCategories/getCategories.ts
