+++ src/entities/Category/api/getCategories/getCategories.types.ts
