+++ src/entities/Category/api/getCategories/index.ts
