+++ src/vite-env.d.ts
@@ -1,6 +1,7 @@
 /// <reference types="vite/client" />
 
 interface ImportMetaEnv {
+  VITE_API_BASE_URL: string;
   VITE_API_KEY: string;
 }
